Dashboard: consolidate the account defaults screen into one folder
Rename wordpress-defaults to preferences-defaults so the screen matches
its route, and follow the index/summary convention used elsewhere in me/
by keeping the summary button alongside the page it links to. The site
dropdown moves back next to its only real consumer.
